feat: add Lua API for prompt input masking.
This commit adds a new Lua API `prompt_mask.set`. This function accepts
a table comprised of integer offsets to string content to insert into
the prompt data at that location pre-rendering. Mask content is not sent
through to the MUD, or added to history data.

The primary usecase for this API is to support adding ANSI colour codes
at predefined offsets to "highlight" words (e.g. based on spellchecking
result).

When the input prompt buffer is changed we clear the prompt mask,
ensuring that the two remain in sync. When a prompt mask is set the
prompt data is freshly rendered using the overlaid mask content.

TODO:
* Test coverage.
* Fix bug when long input line is wrapped.
* ???