net: replace native-tls with rustls.
This commit replaces the dependency used for TLS MUD connections. The
`native-tls` crate is replaced with `rustls`. Rustls has the advantage
of being a pure Rust implementation of TLS 1.2 and TLS 1.3, requiring no
native code except in the underlying *ring* dependency that provides
cryptographic primitives.

Since we're no longer delegating to the system specific TLS
implementation via `native-tls` we must provide our own root certificate
store. For this purpose we use the `webpki-roots` crate. It provides
a reasonable set of default root certificates that closely follows
Mozilla's curated root program.

Blightmud still requires OpenSSL after this commit due to the `git2`
dependency, but it is no longer used when communicating to MUD servers.